Updated Example.Api/Program.cs to net6.0 format and removed Startup class
